+++ src/helpers/index.ts
@@ -22,7 +22,4 @@
  * SOFTWARE.
  */
 
-/**
- * Default context type.
- */
-export type Context = Record<string, unknown>;
+export * from './defaultAuthFn';