+++ src/types/Auth.ts
@@ -25,11 +25,14 @@
 import type { AuthFn } from './AuthFn';
 import type { AuthFnClass } from './AuthFnClass';
 import type { ClassType } from './ClassType';
-import type { Context } from './Context';
 
 /**
  * Auth function or class.
  */
-export type Auth<TContext = Context, TRole = string, TPermission = string> =
+export type Auth<
+  TContext = Record<string, unknown>,
+  TRole = string,
+  TPermission = string
+> =
   | AuthFn<TContext, TRole, TPermission>
   | ClassType<AuthFnClass<TContext, TRole, TPermission>>;