+++ src/types/index.ts
@@ -33,6 +33,5 @@ export * from './ClassType';
 export * from './ClassTypeEmptyConstructor';
 export * from './ContainerGetter';
 export * from './ContainerType';
-export * from './Context';
 export * from './ToArrayStringArgs';
 export * from './ResolverData';