+++ CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Default auth procedure `defaultAuthFn`
+
 ### Changed
 
 - Removed source map for `.d.ts` files
+
+- Removed `Context` type
 
 ## [v0.3.1](https://github.com/carlocorradini/graphql-auth-directive/releases/tag/v0.3.1) - 2022-10-13